+++ main.go
@@ -1,157 +1,70 @@
 package main
 
 import (
+	"errors"
+	"fmt"
 	"log"
 	"net/http"
 	"net/url"
 	"os"
 	"os/signal"
+	"reflect"
 	"strconv"
 	"strings"
 	"syscall"
 
+	"code.cloudfoundry.org/clock"
 	"code.cloudfoundry.org/lager"
+	uaa "code.cloudfoundry.org/uaa-go-client"
+	uaaconf "code.cloudfoundry.org/uaa-go-client/config"
+	credhub "github.com/cloudfoundry-community/go-credhub"
+	"github.com/hashicorp/go-cleanhttp"
 	"github.com/hashicorp/vault/api"
+	"github.com/kelseyhightower/envconfig"
 	"github.com/pivotal-cf/brokerapi"
-)
-
-const (
-	// DefaultListenAddr is the default address unless we get
-	// an override via PORT
-	DefaultListenAddr = ":8000"
-
-	// DefaultServiceID is the default UUID of the services
-	DefaultServiceID = "0654695e-0760-a1d4-1cad-5dd87b75ed99"
-
-	// DefaultVaultAddr is the default address to the Vault cluster.
-	DefaultVaultAddr = "https://127.0.0.1:8200"
-
-	// DefaultServiceName is the name of the service in the marketplace
-	DefaultServiceName = "hashicorp-vault"
-
-	// DefaultServiceDescription is the default service description.
-	DefaultServiceDescription = "HashiCorp Vault Service Broker"
-
-	// DefaultPlanName is the name of our plan, only one supported
-	DefaultPlanName = "shared"
-
-	// DefaultPlanDescription is the default description.
-	DefaultPlanDescription = "Secure access to Vault's storage and transit backends"
 )
 
 func main() {
 	// Setup the logger - intentionally do not log date or time because it will
 	// be prefixed in the log output by CF.
 	logger := log.New(os.Stdout, "", 0)
+	cfLogger := lager.NewLogger("vault-broker")
 
-	// Ensure username and password are present
-	username := os.Getenv("SECURITY_USER_NAME")
-	if username == "" {
-		logger.Fatal("[ERR] missing SECURITY_USER_NAME")
-	}
-	password := os.Getenv("SECURITY_USER_PASSWORD")
-	if password == "" {
-		logger.Fatal("[ERR] missing SECURITY_USER_PASSWORD")
-	}
-
-	// Get a custom GUID
-	serviceID := os.Getenv("SERVICE_ID")
-	if serviceID == "" {
-		serviceID = DefaultServiceID
-	}
-
-	// Get the service name
-	serviceName := os.Getenv("SERVICE_NAME")
-	if serviceName == "" {
-		serviceName = DefaultServiceName
-	}
-
-	// Get the service description
-	serviceDescription := os.Getenv("SERVICE_DESCRIPTION")
-	if serviceDescription == "" {
-		serviceDescription = DefaultServiceDescription
-	}
-
-	// Get the service tags
-	serviceTags := strings.Split(os.Getenv("SERVICE_TAGS"), ",")
-
-	// Get the plan name
-	planName := os.Getenv("PLAN_NAME")
-	if planName == "" {
-		planName = DefaultPlanName
-	}
-
-	// Get the plan description
-	planDescription := os.Getenv("PLAN_DESCRIPTION")
-	if planDescription == "" {
-		planDescription = DefaultPlanDescription
-	}
-
-	// Parse the port
-	port := os.Getenv("PORT")
-	if port == "" {
-		port = DefaultListenAddr
-	} else {
-		if port[0] != ':' {
-			port = ":" + port
-		}
-	}
-
-	// Check for vault address
-	vaultAddr := os.Getenv("VAULT_ADDR")
-	if vaultAddr == "" {
-		vaultAddr = DefaultVaultAddr
-	}
-	os.Setenv("VAULT_ADDR", normalizeAddr(vaultAddr))
-
-	// Get the vault advertise addr
-	vaultAdvertiseAddr := os.Getenv("VAULT_ADVERTISE_ADDR")
-	if vaultAdvertiseAddr == "" {
-		vaultAdvertiseAddr = normalizeAddr(vaultAddr)
-	}
-
-	// Check if renewal is enabled
-	renew := true
-	if s := os.Getenv("VAULT_RENEW"); s != "" {
-		b, err := strconv.ParseBool(s)
-		if err != nil {
-			logger.Fatalf("[ERR] failed to parse VAULT_RENEW: %s", err)
-		}
-		renew = b
-	}
-
-	// Check for vault token
-	if v := os.Getenv("VAULT_TOKEN"); v == "" {
-		logger.Fatal("[ERR] missing VAULT_TOKEN")
+	config, err := parseConfig(cfLogger)
+	if err != nil {
+		logger.Fatal("[ERR] failed to read configuration", err)
 	}
 
 	// Setup the vault client
-	client, err := api.NewClient(nil)
+	vaultClientConfig := api.DefaultConfig()
+	vaultClientConfig.HttpClient = cleanhttp.DefaultClient()
+
+	vaultClient, err := api.NewClient(vaultClientConfig)
 	if err != nil {
-		logger.Fatal("[ERR] failed to create api client", err)
+		logger.Fatal("[ERR] failed to create vault api client", err)
 	}
-	
-	// Check for vault namespace
-	namespace := os.Getenv("VAULT_NAMESPACE")
-	if len(namespace) > 0 {
-		client.SetNamespace(namespace)
+
+	vaultClient.SetAddress(config.VaultAddr)
+	vaultClient.SetToken(config.VaultToken)
+	if config.VaultNamespace != "" {
+		vaultClient.SetNamespace(config.VaultNamespace)
 	}
 
 	// Setup the broker
 	broker := &Broker{
-		log:    logger,
-		client: client,
+		log:         logger,
+		vaultClient: vaultClient,
 
-		serviceID:          serviceID,
-		serviceName:        serviceName,
-		serviceDescription: serviceDescription,
-		serviceTags:        serviceTags,
+		serviceID:          config.ServiceID,
+		serviceName:        config.ServiceName,
+		serviceDescription: config.ServiceDescription,
+		serviceTags:        config.ServiceTags,
 
-		planName:        planName,
-		planDescription: planDescription,
+		planName:        config.PlanName,
+		planDescription: config.PlanDescription,
 
-		vaultAdvertiseAddr: vaultAdvertiseAddr,
-		vaultRenewToken:    renew,
+		vaultAdvertiseAddr: config.VaultAdvertiseAddr,
+		vaultRenewToken:    config.VaultRenew,
 	}
 	if err := broker.Start(); err != nil {
 		logger.Fatalf("[ERR] failed to start broker: %s", err)
@@ -159,18 +72,18 @@ func main() {
 
 	// Parse the broker credentials
 	creds := brokerapi.BrokerCredentials{
-		Username: username,
-		Password: password,
+		Username: config.SecurityUserName,
+		Password: config.SecurityUserPassword,
 	}
 
 	// Setup the HTTP handler
-	handler := brokerapi.New(broker, lager.NewLogger("vault-broker"), creds)
+	handler := brokerapi.New(broker, cfLogger, creds)
 
 	// Listen to incoming connection
 	serverCh := make(chan struct{}, 1)
 	go func() {
-		logger.Printf("[INFO] starting server on %s", port)
-		if err := http.ListenAndServe(port, handler); err != nil {
+		logger.Printf("[INFO] starting server on %s", config.Port)
+		if err := http.ListenAndServe(config.Port, handler); err != nil {
 			logger.Fatalf("[ERR] server exited with: %s", err)
 		}
 		close(serverCh)
@@ -233,4 +146,137 @@ func normalizeAddr(s string) string {
 	u.Path = strings.TrimRight(u.Path, "/") + "/"
 
 	return u.String()
+}
+
+func parseConfig(cfLogger lager.Logger) (*Configuration, error) {
+	config := &Configuration{}
+	if err := envconfig.Process("", config); err != nil {
+		return nil, err
+	}
+	if config.CredhubURL != "" {
+		if err := credhubProcess(cfLogger, "VAULT_SERVICE_BROKER_", config); err != nil {
+			return nil, err
+		}
+	}
+	if err := config.Validate(); err != nil {
+		return nil, err
+	}
+	return config, nil
+}
+
+type Configuration struct {
+	// Required
+	SecurityUserName     string `envconfig:"security_user_name"`
+	SecurityUserPassword string `envconfig:"security_user_password"`
+	VaultToken           string `envconfig:"vault_token"`
+
+	// Optional, for using CredHub
+	CredhubURL                       string `envconfig:"credhub_url"`
+	UAAEndpoint                      string `envconfig:"uaa_endpoint"`
+	UAAClientName                    string `envconfig:"uaa_client_name"`
+	UAAClientSecret                  string `envconfig:"uaa_client_secret"`
+	UAACACerts                       string `envconfig:"uaa_ca_certs"`
+	UAASkipVerification              bool   `envconfig:"uaa_skip_verification"`
+	UAAInsecureAllowAnySigningMethod bool   `envconfig:"uaa_insecure_allow_any_signing_method"`
+
+	// Also optional
+	Port               string   `envconfig:"port" default:":8000"`
+	ServiceID          string   `envconfig:"service_id" default:"0654695e-0760-a1d4-1cad-5dd87b75ed99"`
+	VaultAddr          string   `envconfig:"vault_addr" default:"https://127.0.0.1:8200"`
+	VaultAdvertiseAddr string   `envconfig:"vault_advertise_addr"`
+	VaultNamespace 		string `envconfig:"vault_namespace"`
+	ServiceName        string   `envconfig:"service_name" default:"hashicorp-vault"`
+	ServiceDescription string   `envconfig:"service_description" default:"HashiCorp Vault Service Broker"`
+	PlanName           string   `envconfig:"plan_name" default:"shared"`
+	PlanDescription    string   `envconfig:"plan_description" default:"Secure access to Vault's storage and transit backends"`
+	ServiceTags        []string `envconfig:"service_tags"`
+	VaultRenew         bool     `envconfig:"vault_renew" default:"true"`
+}
+
+func (c *Configuration) Validate() error {
+	// Ensure required parameters were provided
+	if c.SecurityUserName == "" {
+		return errors.New("missing SECURITY_USER_NAME")
+	}
+	if c.SecurityUserPassword == "" {
+		return errors.New("missing SECURITY_USER_PASSWORD")
+	}
+	if c.VaultToken == "" {
+		return errors.New("missing VAULT_TOKEN")
+	}
+
+	// If these values aren't perfect, we can fix them
+	if !strings.HasPrefix(c.Port, ":") {
+		c.Port = ":" + c.Port
+	}
+	if c.VaultAdvertiseAddr == "" {
+		c.VaultAdvertiseAddr = c.VaultAddr
+	}
+	c.VaultAddr = normalizeAddr(c.VaultAddr)
+	c.VaultAdvertiseAddr = normalizeAddr(c.VaultAdvertiseAddr)
+	return nil
+}
+
+// credhubProcess iterates over the names of variables as set in the `envconfig` tag
+// on the Configuration. It prepends them with "prefix" and then looks
+// in Credhub to see if they exist. If they do and they have a value, the Configuration
+// is updated with that value for that field.
+func credhubProcess(cfLogger lager.Logger, prefix string, config *Configuration) error {
+	uaaConf := &uaaconf.Config{
+		ClientName:                    config.UAAClientName,
+		ClientSecret:                  config.UAAClientSecret,
+		UaaEndpoint:                   config.UAAEndpoint,
+		SkipVerification:              config.UAASkipVerification,
+		CACerts:                       config.UAACACerts,
+		InsecureAllowAnySigningMethod: config.UAAInsecureAllowAnySigningMethod,
+	}
+	uaaClient, err := uaa.NewClient(cfLogger, uaaConf, clock.NewClock())
+	if err != nil {
+		return err
+	}
+	credhubClient := credhub.New(config.CredhubURL, credhub.NewUAAAuthClient(cleanhttp.DefaultClient(), uaaClient))
+
+	// Pull the "envconfig" field name from each field and look for it in Credhub
+	configTypeInfo := reflect.TypeOf(*config)
+	settableConfig := reflect.ValueOf(config).Elem()
+
+	for i := 0; i < configTypeInfo.NumField(); i++ {
+		fieldTypeInfo := configTypeInfo.Field(i)
+		credhubName := prefix + strings.ToUpper(fieldTypeInfo.Tag.Get("envconfig"))
+
+		latest, err := credhubClient.GetLatestByName(credhubName)
+		if err != nil && !strings.Contains(strings.ToLower(err.Error()), "not found") {
+			return err
+		}
+		if latest == nil {
+			// This key doesn't exist in Credhub
+			continue
+		}
+		settingValue, ok := latest.Value.(string)
+		if !ok {
+			return fmt.Errorf("we only support credhub values as bash-like string values, but received %s as a %s", credhubName, reflect.TypeOf(latest.Value))
+		}
+		if settingValue == "" {
+			// The value for this key isn't set in Credhub
+			continue
+		}
+
+		// Update the value for this field with Credhub's value
+		settableField := settableConfig.Field(i)
+		switch fieldTypeInfo.Type.Kind() {
+		case reflect.Bool:
+			asBool, err := strconv.ParseBool(settingValue)
+			if err != nil {
+				return fmt.Errorf("error parsing bool %s: %s", credhubName, err)
+			}
+			settableField.SetBool(asBool)
+		case reflect.String:
+			settableField.SetString(settingValue)
+		case reflect.Slice:
+			settableField.Set(reflect.ValueOf(strings.Split(settingValue, ",")))
+		default:
+			return fmt.Errorf("unsupported type of %s for %s", fieldTypeInfo.Type.Kind(), credhubName)
+		}
+	}
+	return nil
 }